clean up cli parsing
