+++ src/cli.rs
@@ -1,14 +1,8 @@
-use std::path::Path;
-
-use crate::picker::get_skim_selection_from_slice;
-use crate::workspace::get_workspace_type_for_path;
+use crate::handler::{
+    handle_existing_session_selection, handle_group_session_selection, handle_workspace_selection,
+};
 use anyhow::Result;
 
-use crate::config::TwmGlobal;
-use crate::matches::find_workspaces_in_dir;
-use crate::tmux::{
-    attach_to_tmux_session, get_tmux_sessions, open_workspace, open_workspace_in_group,
-};
 use clap::Parser;
 
 #[derive(Parser, Default, Debug)]
@@ -39,7 +33,7 @@ pub struct Arguments {
     /// Open the given path as a workspace.
     ///
     /// Using this option does not require that the path be a valid workspace according to your configuration.
-    path: Option<String>,
+    pub path: Option<String>,
 
     #[clap(short, long)]
     /// Force the workspace to be opened with the given name.
@@ -56,39 +50,9 @@ pub struct Arguments {
 pub fn parse() -> Result<()> {
     let args = Arguments::parse();
 
-    if args.existing {
-        let existing_sessions = get_tmux_sessions()?;
-        let session_name =
-            get_skim_selection_from_slice(&existing_sessions, "Select a session to attach to: ")?;
-        attach_to_tmux_session(&session_name)?;
-        Ok(())
-    } else if args.group {
-        let existing_sessions = get_tmux_sessions()?;
-        let group_session_name =
-            get_skim_selection_from_slice(&existing_sessions, "Select a session to group with: ")?;
-        open_workspace_in_group(&group_session_name, &args)?;
-        Ok(())
-    } else {
-        let config = TwmGlobal::load()?;
-        let mut matched_workspace_paths = Vec::<String>::new();
-        // handle a path directly being passed in first
-        let workspace_path = if let Some(path) = args.path.clone() {
-            let path_full = std::fs::canonicalize(path)?;
-            match path_full.to_str() {
-                Some(p) => p.to_owned(),
-                None => anyhow::bail!("Path is not valid UTF-8"),
-            }
-        } else {
-            for dir in &config.search_paths {
-                find_workspaces_in_dir(dir.as_str(), &config, &mut matched_workspace_paths)
-            }
-            get_skim_selection_from_slice(&matched_workspace_paths, "Select a workspace: ")?
-        };
-
-        let workspace_type =
-            get_workspace_type_for_path(Path::new(&workspace_path), &config.workspace_definitions);
-        open_workspace(&workspace_path, workspace_type, &config, &args)?;
-
-        Ok(())
+    match args {
+        Arguments { existing: true, .. } => handle_existing_session_selection(),
+        Arguments { group: true, .. } => handle_group_session_selection(&args),
+        _ => handle_workspace_selection(&args),
     }
 }
